fix (improve) data port definition on netperf
-P sets both ports for tcp_rr, etc.
This has the unintended consequence that running on localhost results in
increased throughput:

```
$ netperf -P0  -l 10 -H localhost -j -t tcp_rr --    -k TRANSACTION_RATE -r 1,1 -b 128
TRANSACTION_RATE=205378.587
$ netperf -P0  -l 10 -H localhost -j -t tcp_rr -- -P 8000   -k TRANSACTION_RATE -r 1,1 -b 128
TRANSACTION_RATE=819730.579
```

Why? my guess is that due to reusing the port the client just reads what
it writes without the data going through the net stack.

Using -P ,XXX sets only the remote port (not the local):

```
$ netperf -P0  -l 10 -H localhost -j -t tcp_rr -- -P ,8000   -k TRANSACTION_RATE -r 1,1 -b 128
[TRANSACTION_RATE=215838.974
```

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>