cmd: Make client affinity default to different
The previous default was none which doesn't enforce any client affinity.
That means tests can produce significantly different results based on
whether the pods are on the same node or not. We likely want to make
sure running the same command twice produces the same results (or at
least the same test environment) so let's remove 'none' and make
'different' the default.

Signed-off-by: Paul Chaignon <paul@cilium.io>